// __generated__/__types__.tsx
export type Maybe<T> = T;
// ...
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  Date: any;
};

export type Analytics = {
  __typename?: 'Analytics';
  google_analytics?: Maybe<Scalars['String']>;
  umami_id?: Maybe<Scalars['String']>;
  simple_analytics?: Maybe<Scalars['String']>;
  plausible_data_domain?: Maybe<Scalars['String']>;
};

export type Author = {
  __typename?: 'Author';
  id: Scalars['Int'];
  email: Scalars['String'];
  username: Scalars['String'];
  name: Scalars['String'];
  social?: Maybe<Social>;
  bio?: Maybe<Scalars['String']>;
  occupation?: Maybe<Scalars['String']>;
  company_name?: Maybe<Scalars['String']>;
  role?: Maybe<Role>;
  permissions?: Maybe<Array<Permissions>>;
  avatar?: Maybe<Scalars['String']>;
  verified?: Maybe<Scalars['Boolean']>;
  accessToken?: Maybe<Scalars['String']>;
  analytics_id?: Maybe<Scalars['Int']>;
  analytics_uuid?: Maybe<Scalars['String']>;
};
// ...
export type Image = {
  __typename?: 'Image';
  src?: Maybe<Scalars['String']>;
  width?: Maybe<Scalars['Int']>;
  height?: Maybe<Scalars['Int']>;
};

export type InputAnalytics = {
  google_analytics?: Maybe<Scalars['String']>;
  umami_id?: Maybe<Scalars['String']>;
  simple_analytics?: Maybe<Scalars['String']>;
  plausible_data_domain?: Maybe<Scalars['String']>;
};
// ...
export type InputImage = {
  height?: Maybe<Scalars['Int']>;
  src: Scalars['String'];
  width?: Maybe<Scalars['Int']>;
};

export type InputNavigation = {
  type: NavigationType;
  slug: Scalars['String'];
  original_name: Scalars['String'];
  label: Scalars['String'];
};
// ...
export type Navigation = {
  __typename?: 'Navigation';
  type: NavigationType;
  slug: Scalars['String'];
  original_name: Scalars['String'];
  label: Scalars['String'];
};

export enum NavigationType {
  Tag = 'tag',
  Page = 'page',
  Custom = 'custom'
}

export enum Permissions {
  ManageOwnPosts = 'MANAGE_OWN_POSTS',
  ReadOnlyPosts = 'READ_ONLY_POSTS',
  ManageAllPosts = 'MANAGE_ALL_POSTS',
  ManageUsers = 'MANAGE_USERS',
  ManageSettings = 'MANAGE_SETTINGS'
}

export type Post = {
  __typename?: 'Post';
  id: Scalars['Int'];
  title: Scalars['String'];
  html?: Maybe<Scalars['String']>;
  html_draft?: Maybe<Scalars['String']>;
  author?: Maybe<Author>;
  excerpt?: Maybe<Scalars['String']>;
  cover_image: Image;
  type: PostTypes;
  status: PostStatusOptions;
  featured: Scalars['Boolean'];
  slug?: Maybe<Scalars['String']>;
  createdAt: Scalars['Date'];
  publishedAt?: Maybe<Scalars['Date']>;
  scheduledAt?: Maybe<Scalars['Date']>;
  updatedAt: Scalars['Date'];
  reading_time?: Maybe<Scalars['String']>;
  tags?: Maybe<Array<Tags>>;
};
// ...
export enum PostStatusOptions {
  Published = 'published',
  Draft = 'draft',
  Trashed = 'trashed'
}

export enum PostTypes {
  Post = 'post',
  Page = 'page'
}
// ...
export type PostsResponse = PostsNode | PostError;
// ...
export enum Role {
  Admin = 'ADMIN',
  Reviewer = 'REVIEWER',
  Reader = 'READER',
  Author = 'AUTHOR'
}

export type Setting = {
  __typename?: 'Setting';
  id: Scalars['Int'];
  site_title: Scalars['String'];
  site_tagline?: Maybe<Scalars['String']>;
  site_email: Scalars['String'];
  site_url: Scalars['String'];
  site_footer?: Maybe<Scalars['String']>;
  site_description?: Maybe<Scalars['String']>;
  subscribe_embed?: Maybe<Scalars['String']>;
  /** @deprecated Field no longer supported */
  social_twitter?: Maybe<Scalars['String']>;
  /** @deprecated Field no longer supported */
  social_facebook?: Maybe<Scalars['String']>;
  /** @deprecated Field no longer supported */
  social_instagram?: Maybe<Scalars['String']>;
  /** @deprecated Field no longer supported */
  social_github?: Maybe<Scalars['String']>;
  display_author_info: Scalars['Boolean'];
  intro_dismissed: Scalars['Boolean'];
  show_about_page?: Maybe<Scalars['Boolean']>;
  show_tags_page?: Maybe<Scalars['Boolean']>;
  cloudinary_key?: Maybe<Scalars['String']>;
  cloudinary_name?: Maybe<Scalars['String']>;
  cloudinary_secret?: Maybe<Scalars['String']>;
  menu: Array<Navigation>;
  css?: Maybe<Scalars['String']>;
  google_analytics?: Maybe<Scalars['String']>;
  analytics?: Maybe<Analytics>;
  theme?: Maybe<Scalars['String']>;
  banner?: Maybe<Image>;
  site_logo?: Maybe<Image>;
  site_favicon?: Maybe<Image>;
  client_token: Scalars['String'];
};

export type SettingError = LetterpadError & {
  __typename?: 'SettingError';
  message: Scalars['String'];
};

export type SettingInputType = {
  site_title?: Maybe<Scalars['String']>;
  site_tagline?: Maybe<Scalars['String']>;
  site_email?: Maybe<Scalars['String']>;
  site_url?: Maybe<Scalars['String']>;
  site_footer?: Maybe<Scalars['String']>;
  site_description?: Maybe<Scalars['String']>;
  subscribe_embed?: Maybe<Scalars['String']>;
  social_twitter?: Maybe<Scalars['String']>;
  social_facebook?: Maybe<Scalars['String']>;
  social_instagram?: Maybe<Scalars['String']>;
  social_github?: Maybe<Scalars['String']>;
  display_author_info?: Maybe<Scalars['Boolean']>;
  cloudinary_key?: Maybe<Scalars['String']>;
  cloudinary_name?: Maybe<Scalars['String']>;
  cloudinary_secret?: Maybe<Scalars['String']>;
  menu?: Maybe<Array<InputNavigation>>;
  analytics?: Maybe<InputAnalytics>;
  css?: Maybe<Scalars['String']>;
  google_analytics?: Maybe<Scalars['String']>;
  theme?: Maybe<Scalars['String']>;
  banner?: Maybe<InputImage>;
  site_logo?: Maybe<InputImage>;
  site_favicon?: Maybe<InputImage>;
  intro_dismissed?: Maybe<Scalars['Boolean']>;
  show_about_page?: Maybe<Scalars['Boolean']>;
  show_tags_page?: Maybe<Scalars['Boolean']>;
};

export type SettingResponse = Setting | SettingError;

export type Social = {
  __typename?: 'Social';
  twitter?: Maybe<Scalars['String']>;
  facebook?: Maybe<Scalars['String']>;
  github?: Maybe<Scalars['String']>;
  instagram?: Maybe<Scalars['String']>;
  linkedin?: Maybe<Scalars['String']>;
};
// ...
export type Tags = {
  __typename?: 'Tags';
  name: Scalars['String'];
  slug: Scalars['String'];
  posts?: Maybe<PostsResponse>;
};